chore(#64): demo, module, entry
