+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/About/Internal/ListItems/MoreApps/MoreAppsBO.swift
@@ -40,33 +40,6 @@ struct MoreAppsAppDetails {
     let storeURL: URL?
 }
 
-// ==================================
-// MARK: - More Apps Apps Plus Mapper
-// ===================================
-
-struct MoreAppsAppsPlusMapper {
-
-    func appsSections(from appsList: AppsPlusListDTO) -> [MoreAppsSection] {
-        appsList.sections.map { appsSection(from: $0) }
-    }
-
-    func appsDetails(from appsList: AppsPlusListDTO) -> [MoreAppsAppDetails] {
-        appsList.apps.map { appDetails(from: $0) }
-    }
-
-    func appsSection(from section: AppsPlusSectionDTO) -> MoreAppsSection {
-        MoreAppsSection(description: section.description,
-                        apps: section.apps.map { appDetails(from: $0) })
-    }
-
-    func appDetails(from details: AppsPlusAppDetailsDTO) -> MoreAppsAppDetails {
-        MoreAppsAppDetails(title: details.title,
-                           iconURL: URL(string: details.iconURL),
-                           description: details.description.isEmpty ? nil : details.description,
-                           storeURL: URL(string: details.storeLink))
-    }
-}
-
 // ============================
 // MARK: - More Apps Repository
 // ============================
