+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Modules/About/Internal/ListItems/MoreApps/MoreAppsView.swift
@@ -1,0 +1,18 @@
+//
+// Software Name: Orange Design System (iOS)
+// SPDX-FileCopyrightText: Copyright (c) 2021 - 2023 Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license.
+//
+
+import SwiftUI
+
+struct MoreAppsView: View {
+
+    let viewModel: MoreAppsViewModel
+
+    var body: some View {
+        Text(°°"modules.about.more_apps.more")
+    }
+}
