+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Resources/Base.lproj/Localizable.strings
@@ -24,3 +24,4 @@
 "modules.about.copyright" = "Orange property. All rights reserved";
 "modules.about.app_news.no_news" = "No App news found in file";
 "modules.about.more_apps.more" = "More Apps will be here";
+"modules.about.apps_recirculation.title" = "More apps";
