+++ OrangeDesignSystemDemo/OrangeDesignSystemDemo/Resources/fr.lproj/Localizable.strings
@@ -73,9 +73,9 @@
 "screens.about.app_information.option_description.feedback" = "Avis";
 "screens.about.app_information.option_description.app_news" = "Actualités de l'app";
 "screens.about.app_information.option_description.legal_information" = "Mentions légales";
+"screens.about.app_information.option_description.more_apps" = "Plus d'apps";
 "screens.about.app_information.option_description.rate" = "Évaluez l'app";
 "screens.about.file_load_failure" = "Impossible de charger le fichier %@.%@";
-"screens.about.apps_recirculation.title" = "Plus d'apps";
 
 // =====================================
 // MARK: - Screens - Guidelines - Colors
